Add rate limiting
Defines different limits for the various paths.

Only uses an in memory rate counter at the moment.
This is easy to circumvent but it's fast and simple enough to provide some protection.

It also provides 3 modes: silent, report and error. This allows us to
change the behaviour without a code change. It allows us to log and not
send an error to the client. This helps us observe and analyse the
traffic, spotting false positives, before actually blocking clients.